Redefine the way methods are patched.
Prior to this methods were cached, then monkey patched, wrapping calls to the
cached methods with logging, then calling them.

This also meant that for test isolation, we had to redefine the methods again,
but without the logging and just passing straight through to the cached
versions.

This was mostly OK, with the exception of how calls to `super` work.

Given the following code

```ruby
def self.configuration=(config)
end

configuration={}

def method_logging_call(*args, **keyword_args, &block)
end
```

Inside the `method_logging_call` for the call relating to `configuration={}`
the values of `*args`, and `**keyword_args` are `[]` and `{}` respectively.

In our method logging code there is no way of inferring whether no argument was
passed or an empty hash was passed.

We could look at function arity and guess. But we still wouldn't know for sure.

This replaces monkey patching with dynamically injecting modules which we can
undefined methods from. Leaving the module tree a mess, but we at least get the
right interpretation of the arguments.